+++ packages/hooks/src/index.ts
@@ -1,8 +1,8 @@
-export * from "@/hooks/use-animated-value";
-export * from "@/hooks/use-copy-to-clipboard";
-export * from "@/hooks/use-event";
-export * from "@/hooks/use-is-mobile";
-export * from "@/hooks/use-media-query";
-export * from "@/hooks/use-mutation-observer";
-export * from "@/hooks/use-pagination";
-export * from "@/hooks/use-state-params";
+export * from "@/use-animated-value";
+export * from "@/use-copy-to-clipboard";
+export * from "@/use-event";
+export * from "@/use-is-mobile";
+export * from "@/use-media-query";
+export * from "@/use-mutation-observer";
+export * from "@/use-pagination";
+export * from "@/use-state-params";
