+++ packages/hooks/src/use-animated-value.test.ts
@@ -1,6 +1,6 @@
 import { act, renderHook } from "@testing-library/react";
 
-import { useAnimatedValue } from "@/hooks/use-animated-value";
+import { useAnimatedValue } from "@/use-animated-value";
 
 describe("useAnimatedValue", () => {
   let originalRequestAnimationFrame: typeof window.requestAnimationFrame;
