+++ packages/hooks/src/use-is-mobile.test.ts
@@ -1,10 +1,10 @@
 import { renderHook } from "@testing-library/react";
 
-import { useIsMobile } from "@/hooks/use-is-mobile";
-import { useMediaQuery } from "@/hooks/use-media-query";
+import { useIsMobile } from "@/use-is-mobile";
+import { useMediaQuery } from "@/use-media-query";
 
 // Mock hook useMediaQuery
-jest.mock("@/hooks/use-media-query", () => ({
+jest.mock("@/use-media-query", () => ({
   useMediaQuery: jest.fn(),
 }));
