+++ packages/hooks/src/use-is-mobile.ts
@@ -1,4 +1,4 @@
-import { useMediaQuery } from "@/hooks/use-media-query";
+import { useMediaQuery } from "@/use-media-query";
 
 /**
  * The breakpoint width in pixels that separates mobile from non-mobile devices.
