+++ packages/hooks/src/use-mutation-observer.test.ts
@@ -1,6 +1,6 @@
 import { renderHook } from "@testing-library/react";
 
-import { useMutationObserver } from "@/hooks/use-mutation-observer";
+import { useMutationObserver } from "@/use-mutation-observer";
 
 describe("useMutationObserver", () => {
   test("should attach a MutationObserver and call the callback on mutations", () => {
