+++ packages/hooks/src/use-pagination.test.ts
@@ -1,6 +1,6 @@
 import { renderHook } from "@testing-library/react";
 
-import { ELLIPSIS, usePagination } from "@/hooks/use-pagination";
+import { ELLIPSIS, usePagination } from "@/use-pagination";
 
 describe("usePagination", () => {
   it("returns an empty array when totalResults is 0", () => {
